+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,9 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+
+18.0.0b40
++++++++
 * Add option `Windows2025` to `--os-sku` for `az aks nodepool add`.
 
 18.0.0b39